requirements.txt
asgiref==3.2.10
Django==2.2
django-rest-framework==0.1.0
djangorestframework==3.11.1
gunicorn==20.0.4
pytz==2020.1
rest-condition==1.0.3
sqlparse==0.3.1
